requirements.txt
# Django and Django REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Production server
gunicorn>=21.0,<23.0

# Static files handling for production
whitenoise>=6.6,<7.0
